# codex/bundles/process-module-architecture-v3/proof/SB34-post-sb33-runtime-reliability-performance-and-e2e-review/swagger-process-routes.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CanDoItAll.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5032/"
    }
  ],
  "paths": {
...
    "/api/processes/launch": {
      "post": {
        "tags": [
          "Processes"
        ],
        "operationId": "LaunchProcess",
        "requestBody": {
          "content": {
            "application/json": "@{schema=}"
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/processes/runs/{runId}/dispatch": {
      "post": {
        "tags": [
          "Processes"
        ],
        "operationId": "DispatchProcessRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": "@{type=string; format=uuid}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}"
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/processes/runs/{runId}/cancel": {
      "post": {
        "tags": [
          "Processes"
        ],
        "operationId": "CancelProcessRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": "@{type=string; format=uuid}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}"
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/processes/runs/{runId}/steps/{stepInstanceId}/rework": {
      "post": {
        "tags": [
          "Processes"
        ],
        "operationId": "RequestProcessStepRework",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": "@{type=string; format=uuid}"
          },
          {
            "name": "stepInstanceId",
            "in": "path",
            "required": true,
            "schema": "@{type=string; format=uuid}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}"
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/processes/runs/{runId}": {
      "get": {
        "tags": [
          "Processes"
        ],
        "operationId": "GetProcessRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": "@{type=string; format=uuid}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/processes/runs/{runId}/history": {
      "get": {
        "tags": [
          "Processes"
        ],
        "operationId": "GetProcessRunHistory",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": "@{type=string; format=uuid}"
          },
          {
            "name": "fromUtc",
            "in": "query",
            "schema": "@{type=string; format=date-time}"
          },
          {
            "name": "toUtc",
            "in": "query",
            "schema": "@{type=string; format=date-time}"
          },
          {
            "name": "take",
            "in": "query",
            "schema": "@{pattern=^-?(?:0|[1-9]\\d*)$; type=System.Object[]; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}